+++ src/com/intellij/idea/plugin/hybris/impex/settings/ImpexSettingsConfigurableProvider.kt
@@ -78,7 +78,7 @@ class ImpexSettingsConfigurableProvider(val project: Project) : ConfigurableProv
             }
             group("Documentation") {
                 row {
-                    documentationEnableCheckBox = checkBox("Enable code folding")
+                    documentationEnableCheckBox = checkBox("Enable documentation")
                         .bindSelected(state.documentation::enabled)
                         .component
                 }